+++ latch_cli/services/init/template/__init__.py
@@ -1,10 +1,3 @@
-"""
-Minimal template workflow to show the structure of a Latch workflow
-
-For a more comprehensive example, see the assemble_and_sort workflow
-For examples on how to use the Latch SDK, see https://docs.latch.bio/examples/workflows_examples.html
-"""
-
 from wf.task import task
 
 from latch.resources.workflow import workflow
@@ -12,17 +5,11 @@ from latch.types.directory import LatchOutputDir
 from latch.types.file import LatchFile
 from latch.types.metadata import LatchAuthor, LatchMetadata, LatchParameter
 
-"""Minimal metadata object - fill in fields with your own values"""
 metadata = LatchMetadata(
-    display_name="CHANGE ME",
-    documentation="CHANGE ME",
+    display_name="WF_NAME",
     author=LatchAuthor(
-        name="CHANGE ME",
-        email="CHANGE ME",
-        github="CHANGE ME",
+        name="AUTHOR_NAME",
     ),
-    repository="CHANGE ME",
-    license="CHANGE ME",
     parameters={
         "input_file": LatchParameter(
             display_name="Input File",
@@ -33,13 +20,11 @@ metadata = LatchMetadata(
             batch_table_column=True,  # Show this parameter in batched mode.
         ),
     },
-    tags=[],
 )
 
 
-# change the name of this function to something more descriptive
 @workflow(metadata)
-def latch_workflow(
+def template_workflow(
     input_file: LatchFile, output_directory: LatchOutputDir
 ) -> LatchFile:
     return task(input_file=input_file, output_directory=output_directory)